bulletin card
completed bulletin card saving UI
