+++ client/src/app/shared/components/employee-register/employee-register.component.html
@@ -187,47 +187,3 @@
 
 <hr>
 
-<form [formGroup]="bulletinForm" (ngSubmit)="addBulletin()">
-    <div class="row">
-        <div class="col-12 col-md-6">
-            <mat-form-field class="w-100" appearance="outline">
-                <mat-label>DEPARTMENT</mat-label>
-                <mat-select id="depId" formControlName="depId" [(value)]="selectedDepartment">
-                    <mat-option *ngFor="let d of departmentDataStore" [value]="d.id">{{d.name}}</mat-option>
-                </mat-select>
-            </mat-form-field>
-        </div>
-        <div class="col-12 col-md-6">
-            <mat-form-field class="w-100" appearance="outline">
-                <mat-label>REDIRECT URL</mat-label>
-                <input matInput id="re-url" formControlName="reUrl" placeholder="Redirect URL (blog)">
-            </mat-form-field>
-        </div>
-    </div>
-
-    <div id="bulletinCard" class="card bulletin-card">
-        <div style="width: 100%; position: relative; margin: 10px 0">
-            <img [src]="titleImg==null?'':onTitleBgSelected()" alt="title image" class="title-image" id="titleImage">
-            <button type="button" class="btn-1 me-2 px-2 title-choose-btn" (click)="chooseBulletinTitle()"><mat-icon>edit</mat-icon></button>
-        </div>
-
-        <input type="text" id="title" formControlName="title" placeholder="Click to add title">
-
-        <input type="text" id="msg" formControlName="msg" placeholder="Click to add description">
-
-        <div class="bulletin-action-btn">
-            <input type="text" id="action" formControlName="action" placeholder="Click to add action">
-        </div>
-
-        <button type="button" class="btn-1 me-2 px-2 bg-choose-btn" (click)="chooseBulletinBg()"><mat-icon>edit</mat-icon></button>
-
-    </div>
-
-    <div>
-        <label for="string-bg">string-bg</label>
-        <input type="text" id="string-bg" formControlName="stringBg">
-    </div>
-
-    <button type="submit">Submit</button>
-</form>
-
