+++ client/src/styles.scss
@@ -187,8 +187,24 @@ body { margin: 0; padding: 0; font-family: Roboto, "Helvetica Neue", sans-serif;
 }
 .mdc-text-field--filled:not(.mdc-text-field--disabled){
   background-color: unset!important;
+  color: var(--color-text-300) !important;
 }
 .mat-form-field-hide-placeholder .mat-mdc-select-placeholder{
+  color: var(--color-text-200) !important;
+}
+.mat-mdc-slide-toggle .mdc-form-field{
+  color: var(--color-text-200) !important;
+}
+.mat-mdc-slide-toggle .mdc-switch--disabled+label{
+  color: var(--color-text-300) !important;
+}
+.mat-mdc-form-field .mat-mdc-floating-label.mdc-floating-label{
+  color: var(--color-text-300) !important;
+}
+.mdc-text-field:not(.mdc-text-field--disabled) .mdc-floating-label{
+  color: var(--color-text-300) !important;
+}
+.mat-mdc-form-field:not(.mat-form-field-no-animations) .mdc-text-field__input{
   color: var(--color-text-200) !important;
 }
